requirements.txt
django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
mssql-django==1.5     # Engine for Azure SQL
pyodbc==5.1.0         # Connection driver for MSSQL
django-storages==1.14 # Connects Django to Azure Blob Storage
azure-storage-blob==12.19.1
redis==5.0.1
celery==5.3.6
django-cors-headers==4.9.0
